qcacmn: Add extra 2g and 5g channels
Add extra 2g channel from 2402 to 2409 with bw 5Mhz
Add extra 2g channel from 2492 to 2499 with bw 5Mhz
Add extra 5g channel from 5010 to 5150 with bw 5-10 Mhz
Add extra 2g channel from 5850 to 5920 with bw 5-10Mhz

CRs-Fixed: 3646145

+++ umac/regulatory/core/src/reg_build_chan_list.c
@@ -1866,7 +1866,7 @@ static void reg_init_chan(struct regulatory_channel *dst_list,
 			  struct wlan_regulatory_psoc_priv_obj *soc_reg)
 {
 	enum channel_enum chan_enum;
-	uint8_t dst_idx;
+	uint16_t dst_idx;
 
 	for (chan_enum = beg_enum; chan_enum <= end_enum; chan_enum++) {
 		dst_idx = chan_enum - dst_idx_adj;
